# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a Complete Ecormmerce server.",
        "title": "Gin Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:9000",
    "basePath": "/",
    "paths": {
...
        "/admin/addproducts": {
            "post": {
                "description": "admin can add product",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Product"
                ],
                "summary": "Admin productadd",
                "operationId": "admin-product add",
                "parameters": [
                    {
                        "type": "string",
                        "description": "name of the  product",
                        "name": "productname",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "price of the  product",
                        "name": "price",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "discription of the  product",
                        "name": "description",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "color of the  product",
                        "name": "color",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "brandID of the  product",
                        "name": "brandID",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "stock of the  product",
                        "name": "stock",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "cotogeryID of the  product",
                        "name": "catogoryID",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "sizeID of the  product",
                        "name": "sizeID",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Upload a product image",
                        "name": "image",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "discount of the  product",
                        "name": "discount",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "brandDiscount of the  product",
                        "name": "BrandDiscount",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/admin/applydiscount": {
            "put": {
                "description": "admin can discount product",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Product"
                ],
                "summary": "Admin product discount",
                "operationId": "admin-product-discount",
                "parameters": [
                    {
                        "description": "brand discount",
                        "name": "Brand",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.Brand"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/admin/deleteproducts/": {
            "delete": {
                "description": "admin can delete product",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Product"
                ],
                "summary": "Admin product delete",
                "operationId": "admin-product delete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id of the  product",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/admin/editproducts": {
            "put": {
                "description": "admin can edit product",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Product"
                ],
                "summary": "Admin product edit",
                "operationId": "admin-product edit",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id of the  product",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "edit product data",
                        "name": "EditProductData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.EditProductsData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
        "/admin/getcategory": {
            "get": {
                "description": "admin can view productlist",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Product"
                ],
                "summary": "Admin productlist viewer",
                "operationId": "admin-productlist-view",
                "parameters": [
                    {
                        "type": "string",
                        "description": "brand of the  product",
                        "name": "brandsearch",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "catogery of the  product",
                        "name": "categorysearch",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "size of the  product",
                        "name": "sizesearch",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
        "/user/get-productbyid": {
            "get": {
                "description": "user can view product with with id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users Product"
                ],
                "summary": "User product viewer",
                "operationId": "user-product-view with id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id of the  product",
                        "name": "id",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
        "/user/show-product-id": {
            "get": {
                "description": "user can view productid with with name",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users Product"
                ],
                "summary": "User productid viewer",
                "operationId": "user-productid-view",
                "parameters": [
                    {
                        "type": "string",
                        "description": "name of the  product",
                        "name": "product-name",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/user/view-products": {
            "get": {
                "description": "user can view product with with id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users Product"
                ],
                "summary": "User product viewer with search",
                "operationId": "user-product-view with search",
                "parameters": [
                    {
                        "type": "string",
                        "description": "searched data of the  product",
                        "name": "search",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "sort products",
                        "name": "sort",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.Brand": {
            "type": "object",
            "properties": {
                "brand_id": {
                    "type": "integer"
                },
                "discount": {
                    "type": "integer"
                }
            }
        },
        "controllers.EditProductsData": {
            "type": "object",
            "properties": {
                "brand": {
                    "type": "string"
                },
                "color": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "productName": {
                    "type": "string"
                }
            }
        },
...
    }
}